rename signup config
